feat: add a label and a placeholder to every field
